+++ sdk/release_notes.md
@@ -10,6 +10,7 @@
 - Updated `Microsoft.Azure.Functions.Worker.Sdk.Generators` reference to 1.3.4.
 - Setting _ToolingSuffix for TargetFrameworkVersion v9.0
 - Adding support for SDK container builds with Functions base images
+- Removed the default value for HttpStatusCode in WriteAsJsonAsync (#2720)
 
 ### Microsoft.Azure.Functions.Worker.Sdk.Generators <version>